finished basic superChicken

+++ ecs/predict.h
@@ -3,6 +3,7 @@
 
 #include <fstream>
 #include <iostream>
+#include <float.h>
 #include <vector>
 #include <math.h>
 #include <string.h>
@@ -83,9 +84,9 @@ void predict_server(char * info[MAX_INFO_NUM], char * data[MAX_DATA_NUM], int da
 void loadInfo(char * info[MAX_INFO_NUM], phyServerInfo &target);
 void sortFlavorOrderByOptimizationTarget(phyServerInfo &target);
 void quickSortMinToMax(int left, int right, int * array);
-void quickSortMinToMax(int left, int right, double *array , int * index);
+void quickSortMinToMax(int left, int right, int *array , int * index);
 void quickSortMaxToMin(int left, int right, int * array);
-void quickSortMaxToMin(int left, int right, double *array , int * index);
+void quickSortMaxToMin(int left, int right, int *array , int * index);
 void loadTrainDataToVector(vector<trainData> &target, int daysCount, char * data[MAX_DATA_NUM], int dataLineCount, phyServerInfo &serverInfo);
 char * charToNum(char * str, int &target);
 void numToDate(int num, date &target);
